Time: 3 ms (94.03%) | Memory: 62.6 MB (10.43%) - LeetSync

+++ 2432-number-of-zero-filled-subarrays/number-of-zero-filled-subarrays.java
@@ -1,7 +1,7 @@
 class Solution {
     public long zeroFilledSubarray(int[] nums) {
         long ans=0;
-        long cnt=1;
+        int cnt=1;
         for(int num:nums){
             if(num==0){
                 ans+=cnt;
